+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add base class for library-defined exceptions (as part of [EXT-15](https://github.com/isnullxbh/ext/issues/15)).
+
 ## [0.1.3] - 2021-05-27
 
 ### Added